refactor(canvas): extract CSS filter parser into skia_canvas_filter.cpp
skia_canvas.cpp is a long-lived merge-conflict hot spot. The CSS
`filter` / Canvas2D `ctx.filter` parsing cluster — skip_ws,
parse_filter_arg{,s,_length}, parse_css_color_to_skcolor, the eight
SkColorMatrix builders, compose, color_matrix_filter, and
parse_filter_chain — is a self-contained set of pure free functions
with a single cross-TU entry point (parse_filter_chain, called once by
SkiaCanvas::set_filter). Moving it into a sibling TU means
filter-parsing work no longer recompiles the whole file. Mirrors the
prior skia_canvas_text.cpp / skia_canvas_mask.cpp / gradients splits.

- New skia_canvas_filter.cpp holds the cluster, PULP_HAS_SKIA-gated,
  with its own clean #ifdef guard and the Skia includes the code needs
  (SkColor, SkColorFilter, SkColorMatrix, SkImageFilters).
- New skia_canvas_filter.hpp declares only parse_filter_chain (the
  other cluster functions stay file-local). Deliberately not
  skia_canvas_internal.hpp: skia_canvas.cpp keeps its own `static`
  to_sk_color4f, so pulling in the inline-defined header would be a
  redefinition conflict.
- parse_filter_chain un-`static`'d for external linkage.
- skia_canvas.cpp drops 429 lines (1879 -> 1450) and gains the include.

Verified with a Skia-linked Debug build: pulp-canvas builds clean;
pulp-test-canvas passes (1375 assertions / 43 cases) and
pulp-test-canvas2d-shim passes (167 assertions / 53 cases). Pure
mechanical move — no behavior change.
