+++ core/canvas/CMakeLists.txt
@@ -86,6 +86,7 @@ if(PULP_HAS_SKIA)
     target_sources(pulp-canvas PRIVATE
         src/skia_canvas.cpp
         src/skia_canvas_text.cpp
+        src/skia_canvas_filter.cpp
         src/skia_canvas_mask.cpp
         src/skia_canvas_gradients.cpp
         src/skia_canvas_box_shadow.cpp
